+++ clean_python/s3/s3_bucket_options.py
@@ -1,3 +1,5 @@
+from typing import Literal
+
 from clean_python import ValueObject
 
 __all__ = ["S3BucketOptions"]
@@ -9,3 +11,6 @@ class S3BucketOptions(ValueObject):
     secret_key: str
     bucket: str
     region: str | None = None
+    addressing_style: Literal[
+        "virtual", "path"
+    ] = "virtual"  # path is deprecated for AWS S3